Drop `yum_requirements.txt`
The only requirement that was needed from `yum` was a Fortran compiler
when using the legacy compilers. However that has long since been
included in the container to improve compatibility between the legacy
and new toolchains. Hence this drops the `devtoolset-2-gcc-gfortran`
requirement from `yum` and for that matter the `yum_requirements.txt`
file entirely. Should also fix a build issue in the new container where
using `yum` currently does not work correctly. Thus allowing us to add
the new compilers to this feedstock.